code/requirements.txt
fastapi
uvicorn
sqlalchemy
aiomysql
pandas
openpyxl
python-multipart
pydantic
python-dotenv